[devscout] introduce SignatureRecordTemplate
Using `SignatureRecordImpl` to represent templates is problematic as it forces users to write nonsensical templates that look like DKIM signatures but are not valid ones.

It also forced the split of the creation and the validation only to support the construction of a specific error case in DKIMVerifier.

The template relaxes the constraints of the record implementation by not requiring the following tags :
 - a
 - b
 - bh

 The template has a method to build a valid signature record when provided :
 - a SigningAlgorithm
 - a HashMethod
 - a body hash
 - a signature

This will be used to further refine JDKIM apis in a later proposal.